+++ src/utils/runCode.js
@@ -1,0 +1,28 @@
+import _ from 'underscore';
+import sinon from 'sinon';
+import vm from 'vm';
+import spy from './spy';
+import getPatchedTape from './tape';
+
+const runCode = (code) => {
+  const test = getPatchedTape();
+  // should hijack setTimeout before pass to sandbox
+  const clock = sinon.useFakeTimers();
+  const sandbox = {
+    setTimeout: window.setTimeout, // need to be passed also...
+    console,
+    sinon,   
+    describe: test,
+    test,
+    clock,
+    spy
+  };
+
+  const script = new vm.Script(code);
+  const context = new vm.createContext(sandbox);
+  script.runInContext(context);
+  clock.restore();
+}
+const debouncedRunCode = _.debounce(runCode, 200);
+
+export default debouncedRunCode ;
